% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNecessaryCorrectionTypeExample.R
\docType{data}
\name{DESKEW_TOOL}
\alias{DESKEW_TOOL}
\title{Value for Deskew tool of Portada}
\format{
An object of class \code{character} of length 1.
}
\usage{
DESKEW_TOOL
}
\description{
This constants represents the Deskew tool of PORTADA project.
}
\keyword{datasets}
