% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNecessaryCorrectionTypeExample.R
\docType{data}
\name{DEWARP_TOOL}
\alias{DEWARP_TOOL}
\title{Value for Dewarp tool of Portada}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEWARP_TOOL
}
\description{
This constants represents the Dewarp tool of PORTADA project.
}
\keyword{datasets}
