% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNecessaryCorrectionTypeExample.R
\docType{data}
\name{FIX_BACK_TRANSPARENCY_TOOL}
\alias{FIX_BACK_TRANSPARENCY_TOOL}
\title{Value for Fix back transparency tool of Portada}
\format{
An object of class \code{character} of length 1.
}
\usage{
FIX_BACK_TRANSPARENCY_TOOL
}
\description{
This constants represents the Fix back transparency tool of PORTADA project.
}
\keyword{datasets}
