% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNecessaryCorrectionTypeExample.R
\name{getNecessaryCorrectionType}
\alias{getNecessaryCorrectionType}
\title{getNecessaryCorrectionType}
\usage{
getNecessaryCorrectionType(daijson)
}
\arguments{
\item{daijson}{Is a string containing the report of Document AI next to aply
an OCR process to image. The text conetent has teh JSON format used by
Document AI}
}
\value{
an array with all the constant values of the tools necessary to
troubleshoot the image that caused the JSON received by parameter.
}
\description{
This function analyses the problems detected by Document AI applying an OCR
process and searches what tool of PORTADA project image treatment can fix it.
}
