% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDif.R
\name{testDif}
\alias{testDif}
\title{testDif}
\usage{
testDif(n1, n2)
}
\arguments{
\item{n1}{Is a number to substract the secont parameter}

\item{n2}{Is a number to sum to the first parameter}
}
\value{
A number with value equals to n1+n2.
}
\description{
This is a function to test the building of packages in R. This is a single
function that return the difference of the parameters.
}
