% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSum.R
\name{testSum}
\alias{testSum}
\title{testSum}
\usage{
testSum(n1, n2)
}
\arguments{
\item{n1}{Is a number to sum to teh secont parameter}

\item{n2}{Is a number to sum to the first parameter}
}
\value{
A number with value equals to n1+n2.
}
\description{
This is a function to test the building of packages in R. This is a single
function that return the sum of the parameters.
}
